Extract local files on the fly if possible
