+++ src/datasets/features/image.py
@@ -10,9 +10,8 @@ import pyarrow as pa
 
 from .. import config
 from ..download.download_config import DownloadConfig
-from ..download.streaming_download_manager import xopen
 from ..table import array_cast
-from ..utils.file_utils import is_local_path
+from ..utils.file_utils import is_local_path, xopen
 from ..utils.py_utils import first_non_null_value, no_op_if_value_is_null, string_to_dict
