+++ src/datasets/packaged_modules/text/text.py
@@ -50,6 +50,7 @@ class Text(datasets.ArrowBasedBuilder):
         """
         if not self.config.data_files:
             raise ValueError(f"At least one data file must be specified, but got data_files={self.config.data_files}")
+        dl_manager.download_config.extract_on_the_fly = True
         data_files = dl_manager.download_and_extract(self.config.data_files)
         if isinstance(data_files, (str, list, tuple)):
             files = data_files
